Create Group Type Contents
This temporarily places a group, deletes it and then rolls it back, allowing you to see what is in a type and compare it against placed groups with missing elements to establish what is missing. It also temporarily places some walls to ensure all placed rooms are bound... otherwise there were incorrect results.